khal-wrapper: Support subaliases ("khal-foo-bar")
For the correct placement of subcommand arguments (which khalr places between the "foo" and "bar" parts, stupid as it is: "khal-wrapper foo -a Resilient bar"), we need to redo the khalArg parsing and collect khalCommand first.
The remainder is mostly my git-wrapper implementation, with the added twist that both "foo-bar" and "foo" "bar" need to be passed to pagerDisableCheck to be able to handle both.