ENH: DWIM: Allow customization of overall and quadrant precision
And use hours for the {day,week} summary and working days for the quadrants.
I want the total in hours (to see how close to 40 h/week I'm) and working days make more sense than 24 h in the quadrants (where I don't want to round to hours as that would loose the quarter-hour precision).